[move] 네비게이션 그래프에 따른 응집화를 위한 파일 이동
