+++ app/src/main/java/com/jeju/nanaland/ui/main/MainScreen.kt
@@ -44,7 +44,7 @@ import com.jeju.nanaland.ui.main.home.HomeScreen
 import com.jeju.nanaland.ui.main.home.HomeViewModel
 import com.jeju.nanaland.ui.main.home.search.SearchViewModel
 import com.jeju.nanaland.ui.nanapick.NanaPickListScreen
-import com.jeju.nanaland.ui.profile.screen.ProfileScreen
+import com.jeju.nanaland.ui.profile.root.ProfileScreen
 import com.jeju.nanaland.ui.theme.NanaLandTheme
 import com.jeju.nanaland.ui.theme.caption02SemiBold
 import com.jeju.nanaland.ui.theme.getColor
