+++ app/src/main/java/com/jeju/nanaland/ui/profile/ProfileModificationScreenRoute.kt
@@ -1,10 +1,10 @@
-package com.jeju.nanaland.ui.navigation
+package com.jeju.nanaland.ui.profile
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.jeju.nanaland.globalvalue.constant.ROUTE_PROFILE_UPDATE
-import com.jeju.nanaland.ui.profileupdate.ProfileUpdateScreen
+import com.jeju.nanaland.ui.profile.profileupdate.ProfileUpdateScreen
 
 fun NavGraphBuilder.profileUpdateScreen(navController: NavController) = composable(route = ROUTE_PROFILE_UPDATE) {
     ProfileUpdateScreen(
