+++ app/src/main/java/com/jeju/nanaland/ui/profile/root/component/parts/ProfileReviewRow.kt
@@ -1,4 +1,4 @@
-package com.jeju.nanaland.ui.profile.component.parts
+package com.jeju.nanaland.ui.profile.root.component.parts
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
