+++ app/src/main/java/com/jeju/nanaland/ui/profile/root/component/parts/ProfileScreenMoreInfoPart.kt
@@ -1,4 +1,4 @@
-package com.jeju.nanaland.ui.profile.component.parts
+package com.jeju.nanaland.ui.profile.root.component.parts
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.border
